fix leave group

+++ iris/src/main/java/ir/mohaymen/iris/subscription/SubscriptionController.java
@@ -28,6 +28,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.text.MessageFormat;
 import java.time.Instant;
 import java.util.*;
 
@@ -79,24 +80,26 @@ public class SubscriptionController extends BaseController {
     @DeleteMapping("/delete-sub/{subId}")
     public ResponseEntity<?> leaveGroupBySubId(@PathVariable Long subId) throws Exception {
         Subscription subscription = subscriptionService.getSubscriptionBySubscriptionId(subId);
+        if (subscription.getUser().getUserId() != getUserByToken().getUserId() && (!subscription.getPermissions().contains(Permission.ADD_USER) ||!subscription.getPermissions().contains(Permission.ADMIN) ))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         if (subscription.getChat().getChatType() == ChatType.PV) {
             chatService.deleteById(subscription.getChat().getChatId());
             return ResponseEntity.ok("PV deleted.");
         }
         subscriptionService.deleteById(subId);
-        return ResponseEntity.ok("you have left the group!");
+        return ResponseEntity.ok(MessageFormat.format("{0} have left the group!" , subId.toString()));
     }
-    @DeleteMapping("/delete-sub/{chatId}")
-    public ResponseEntity<?> leaveGroupByChatId(@PathVariable Long chatId) throws Exception {
-        Chat chat = new Chat(); chat.setChatId(chatId); chat.setChatType(ChatType.PV); chat.setCreatedAt(Instant.now());
-        Subscription subscription = subscriptionService.getSubscriptionByChatAndUser( chat, getUserByToken());
-        if (subscription.getChat().getChatType() == ChatType.PV) {
-            chatService.deleteById(subscription.getChat().getChatId());
-            return ResponseEntity.ok("PV deleted.");
-        }
-        subscriptionService.deleteById(subscription.getSubId());
-        return ResponseEntity.ok("you have left the group!");
-    }
+//    @DeleteMapping("/delete-sub/{chatId}")
+//    public ResponseEntity<?> leaveGroupByChatId(@PathVariable Long chatId) throws Exception {
+//        Chat chat = new Chat(); chat.setChatId(chatId); chat.setChatType(ChatType.PV); chat.setCreatedAt(Instant.now());
+//        Subscription subscription = subscriptionService.getSubscriptionByChatAndUser( chat, getUserByToken());
+//        if (subscription.getChat().getChatType() == ChatType.PV) {
+//            chatService.deleteById(subscription.getChat().getChatId());
+//            return ResponseEntity.ok("PV deleted.");
+//        }
+//        subscriptionService.deleteById(subscription.getSubId());
+//        return ResponseEntity.ok("you have left the group!");
+//    }
     @GetMapping("/chat-subs/{id}")
     public ResponseEntity<List<SubDto>> subsOfOneChat(@PathVariable Long id) {
         List<SubDto> subDtoList = new ArrayList<>();
